Remove enableOverlay and disableOverlay tests
... as these are covered by enableAndDisableOverlay and the
enableOverlay test can cause failures in other tests when it
is run without the disableOverlay test running immediately
afterwards. In this case, the overlay will be left enabled
which is unexpected for other tests which take screen
snapshots. This case occurs in presubmit/postsubmit testing
when tests are sharded across multiple devices.

Bug: b/242755851
Test: SurfaceFlinger_test RefreshRateOverlayTest.enableOverlay
      then
      SurfaceFlinger_test LayerRenderTypeTransactionTest.SetPositionOutOfBounds_BufferQueue/1
Change-Id: I8c0bf93bb314b27a9be180af19b15eed1bb574fe
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>